requirements.txt
aiosmtplib==4.0.0
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
attrs==25.1.0
bcrypt==3.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.8
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
loguru==0.7.3
Mako==1.3.9
MarkupSafe==3.0.2
outcome==1.3.0.post0
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
PySocks==1.7.1
pytest==8.3.4
pytest-asyncio==0.25.3
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2025.1
requests==2.32.3
rsa==4.9
selenium==4.28.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.38
starlette==0.45.3
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
urllib3==2.3.0
uvicorn==0.34.0
webdriver-manager==4.0.2
websocket-client==1.8.0
win32_setctime==1.2.0
wsproto==1.2.0
